fix: userController에서 scheduleController로 이동
